+++ internal/datanode/syncmgr/taskv2_test.go
@@ -41,6 +41,7 @@ import (
 	"github.com/milvus-io/milvus/internal/datanode/metacache"
 	"github.com/milvus-io/milvus/internal/proto/datapb"
 	"github.com/milvus-io/milvus/internal/storage"
+	"github.com/milvus-io/milvus/internal/util/typeutil"
 	"github.com/milvus-io/milvus/pkg/common"
 	"github.com/milvus-io/milvus/pkg/log"
 	"github.com/milvus-io/milvus/pkg/util/paramtable"
@@ -94,7 +95,7 @@ func (s *SyncTaskSuiteV2) SetupSuite() {
 		},
 	}
 
-	arrowSchema, err := metacache.ConvertToArrowSchema(s.schema.Fields)
+	arrowSchema, err := typeutil.ConvertToArrowSchema(s.schema.Fields)
 	s.NoError(err)
 	s.arrowSchema = arrowSchema
 }
