+++ internal/datanode/writebuffer/bf_write_buffer_test.go
@@ -22,6 +22,7 @@ import (
 	"github.com/milvus-io/milvus/internal/proto/datapb"
 	"github.com/milvus-io/milvus/internal/querycoordv2/params"
 	"github.com/milvus-io/milvus/internal/storage"
+	"github.com/milvus-io/milvus/internal/util/typeutil"
 	"github.com/milvus-io/milvus/pkg/common"
 	"github.com/milvus-io/milvus/pkg/mq/msgstream"
 	"github.com/milvus-io/milvus/pkg/util/paramtable"
@@ -206,7 +207,7 @@ func (s *BFWriteBufferSuite) TestBufferDataWithStorageV2() {
 	params.Params.CommonCfg.EnableStorageV2.SwapTempValue("true")
 	params.Params.CommonCfg.StorageScheme.SwapTempValue("file")
 	tmpDir := s.T().TempDir()
-	arrowSchema, err := metacache.ConvertToArrowSchema(s.collSchema.Fields)
+	arrowSchema, err := typeutil.ConvertToArrowSchema(s.collSchema.Fields)
 	s.Require().NoError(err)
 	space, err := milvus_storage.Open(fmt.Sprintf("file:///%s", tmpDir), options.NewSpaceOptionBuilder().
 		SetSchema(schema.NewSchema(arrowSchema, &schema.SchemaOptions{
@@ -235,7 +236,7 @@ func (s *BFWriteBufferSuite) TestAutoSyncWithStorageV2() {
 	params.Params.CommonCfg.EnableStorageV2.SwapTempValue("true")
 	paramtable.Get().Save(paramtable.Get().DataNodeCfg.FlushInsertBufferSize.Key, "1")
 	tmpDir := s.T().TempDir()
-	arrowSchema, err := metacache.ConvertToArrowSchema(s.collSchema.Fields)
+	arrowSchema, err := typeutil.ConvertToArrowSchema(s.collSchema.Fields)
 	s.Require().NoError(err)
 
 	space, err := milvus_storage.Open(fmt.Sprintf("file:///%s", tmpDir), options.NewSpaceOptionBuilder().
